dataplacement: increase max data placement IDs to 128
Some users have requested the ability to test a larger number of
placement IDs in a single job. Bump the max placement IDs to 128.
Change the type to 16 bits to reduce the amount of space these
additional IDs will consume.

Also bump the server version for this change.

+++ thread_options.h
@@ -394,7 +394,7 @@ struct thread_options {
 	unsigned int fdp;
 	unsigned int dp_type;
 	unsigned int dp_id_select;
-	unsigned int dp_ids[FIO_MAX_DP_IDS];
+	uint16_t dp_ids[FIO_MAX_DP_IDS];
 	unsigned int dp_nr_ids;
 	char *dp_scheme_file;
 
@@ -712,7 +712,7 @@ struct thread_options_pack {
 	uint32_t fdp;
 	uint32_t dp_type;
 	uint32_t dp_id_select;
-	uint32_t dp_ids[FIO_MAX_DP_IDS];
+	uint16_t dp_ids[FIO_MAX_DP_IDS];
 	uint32_t dp_nr_ids;
 	uint8_t dp_scheme_file[FIO_TOP_STR_MAX];
